resolving issues on homepage PR
