+++ src/components/Heading.jsx
@@ -1,40 +1,11 @@
-import { Box, Typography, createTheme, ThemeProvider } from "@mui/material";
+import { Box, Typography } from "@mui/material";
 
-const fontTheme = createTheme({
-    components: {
-        MuiTypography: {
-            variants: [
-                {
-                    props: {
-                        variant: "h4",
-                    },
-                    style: {
-                        fontSize: 49,
-                        fontWeight: 700,
-                    },
-                },
-                {
-                    props: {
-                        variant: "h6",
-                    },
-                    style: {
-                        fontSize: 34,
-                        fontWeight: 500,
-                    },
-                },
-            ],
-        },
-    },
-});
-
-export const Heading = ({ title, size }) => {
+export const Heading = ({ title, size, sx }) => {
     return (
-        <Box sx={{ mt: 0 }}>
-            <ThemeProvider theme={fontTheme}>
-                <Typography variant={size} sx={{ color: "tertiary.main", marginBottom: "24px" }}>
-                    {title}
-                </Typography>
-            </ThemeProvider>
+        <Box sx={{ mt: 0, color: "palette.tertiary.main", ...sx }}>
+            <Typography variant={size}>
+                {title}
+            </Typography>
         </Box>
     );
 };
